gets rid of the while(true) loop
there is no point to this. totalWins and totalLosses get reset each time, so it isn;t useful for running it loads of times and then  seeing what an overall pattern is. It is no more useful typing something other than q and pressing enter for it to go again, than it is to just ... run it again. And it also causes problems trying to run it multiple times for an automated test. It is better to have the automated test run it lots of times and calculate the result, than it is to sit there and pr3ess keys to make it run again, and write down each result on paper, then average them all.